refactor: split GLM speculative verifier
